Move color items to params package
